add a test for dfc matching in text import
still working out what standards my tests should be configured with,
so this commit is a bit hacky.

the problem here is that we're in an awkward state with the API
partially using RTK query to store backend data, which we need to mock
with MSW in tests, and partially using async thunks to store backend
data, which we can get away with preloading the state for.

i think the ideal setup here is for all tests to mock the backend fully
with MSW irrespective of whether the implementation uses RTK query
or async thunks. will address this in a future commit.

also didn't realise that we need to polyfill the standard `fetch`
function in our test environment. this is necessary because jest runs
in node.js and node.js doesn't currently include an implementation
of `fetch`. handled by the `whatwg-fetch` library.