resolve issue where brave isn't correctly picked up on linux
webdriver manager wasn't respecting the `chrome_type` of brave
and trying to run the tool with chrome instead.

it turns out that brave support was added to webdriver-manager in 3.5.4
https://github.com/SergeyPirogov/webdriver_manager/releases/tag/v3.5.4

upgrading to the newest webdriver manager version here.

brave support may have regressed when i pinned an older version of
webdriver-manager because of broken logging. not completely sure :/

should resolve https://github.com/chilli-axe/mpc-autofill/issues/142

will need to test on macos and windows before including in a release.