fix the race condition in this test
i *think* what's going on here is the test will occasionally
add cards to the project before the mocked network call for
retrieving double-faced cards has completed.

my solution here is to make our various card import methods aware of
the state of that request and only allow submitting the card list
when that request is no longer in-flight.

then, our tests now explicitly wait until the import method is ready
to be submitted.