keep track of cards in dictionaries keyed by their IDs and merge... ...slots of identical cards across combined projects
this refactor allows for an enhancement to image collection combination
- we can now handle the case where an image is present in both
collections by taking the union of their slots.

for example, say you combine two projects which share the same cardback
- after this change, the tool will continue to recognise it as a single
cardback, so it'll choose "same images" and fill the cardback into one
slot; previously, it would have treated the cardbacks of each project
as distinct, then would have chosen "different cardbacks" which is way
slower and much more painful.

bit painful to update all these tests like this to be honest!
but i'm in utter shock that they all passed on the first try :0000